+++ src/Model/Places/WareHouse.java
@@ -19,7 +19,7 @@ public class WareHouse extends Element {
     }
 
     @Override
-    public void move(double finalX, double finalY) {
+    public void move(int finalX, int finalY) {
         //nothing
     }
