+++ src/Model/Transportation/Truck.java
@@ -46,7 +46,7 @@ public class Truck extends Element {
     }
 
     @Override
-    public void move(double finalX, double finalY) {
+    public void move(int finalX, int finalY) {
         //
     }
